Finish lecture 3 (merging) code files.
Create code and lecture files for lecture 3 on merging with the dplyr
package. Code is split into four files, with the first mirroring
simple examples from the lecture slides, the second showing the
creation and merging of admission-level and patient-level data sets, a
third showing how to merge in the Charges files, and the fourth
showing how to merge in the cost-to-charge ratio files.

Videos for lecture and the second R script have been recorded. The
second R script has '-in-class' and 'in-class-actual' files added to
the lecture 3 directory. The same needs to be done for the third and
fourth scripts (it's not necessary for the first).