Define new 'coder' functions and simple tests
Define new functions for a 'coder.R' script that students can `source`
for access to helper functions in assigning indicator and date
variables for diagnoses and procedures. The functions use
`dplyr::case_when` rather than the original (much slower) versions
that used `dplyr::rowwise`/`dplyr::c_across`, but unfortunately we can
only define one variable at a time; hence, the functions take data
sets as their first argument and are named, e.g., `mutate_flag_dx` as
a hint that they should be run as separate steps in a chain of
function calls. It may be possible to simplify them using the
meta-programming features offered by the rlang package, but it's not
immediately clear how that would work.

Write crude tests of the new functions.