Test `apply`-based approach to flagging dx codes
Write a version of `dx_flag` suitable for use as an input to `apply`
to investigate whether this could be a reasonable alternative to the
`mutate`-based approach settled on over the last week.

A test supports the use of the `mutate`-based functions. Firstly, it's
not clear that `apply` would be any easier for students to
understand. Secondly, the code using `apply` took substantially longer
to run, suggesting it may not even be an improvement over
`rowwise`/`c_across`, whose inefficiencies prompted the rewrite.